refactor(perf): move cubesandbox_clean.py into business/ directory
Give the business module a proper home — business/ now holds all
CubeSandbox platform operations (cleanup, script registration, etc.).
Future modules (e.g. template management, volume provisioning) fit
naturally here without polluting the perf root.
