+++ tests/perf/business/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2026 Tencent Inc.
+# SPDX-License-Identifier: Apache-2.0
+"""业务模块 — CubeSandbox 平台操作的封装。
+
+压测后资源清理、脚本导入等平台操作都放在这里，
+通过 SDK 调用底层 API（不拼裸 HTTP）。
+"""
